added and fix stories for general and blockchain page
